+++ example/lib/models/user.dart
@@ -2,6 +2,7 @@ import 'package:flutter_data/flutter_data.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import 'adapters.dart';
+import 'task.dart';
 
 part 'user.g.dart';
 
@@ -9,9 +10,10 @@ part 'user.g.dart';
 @DataRepository([JSONServerAdapter])
 class User with DataModel<User> {
   @override
-  final int id;
+  final int? id;
   final String name;
-  final String email;
+  late final HasMany<Task> tasks;
 
-  User({required this.id, required this.name, required this.email});
+  User({this.id, required this.name, HasMany<Task>? tasks})
+      : tasks = tasks ?? HasMany();
 }