+++ lib/src/model/relationship/relationship.dart
@@ -3,7 +3,7 @@ part of flutter_data;
 /// A `Set` that models a relationship between one or more [DataModel] objects
 /// and their a [DataModel] owner. Backed by a [GraphNotifier].
 abstract class Relationship<E extends DataModel<E>, N>
-    with SetMixin<E>, _Lifecycle {
+    with EquatableMixin, _Lifecycle {
   @protected
   Relationship([Set<E>? models])
       : _uninitializedKeys = {},
@@ -92,7 +92,6 @@ abstract class Relationship<E extends DataModel<E>, N>
   /// Add a [value] to this [Relationship]
   ///
   /// Attempting to add an existing [value] has no effect as this is a [Set]
-  @override
   bool add(E value, {bool notify = true}) {
     if (contains(value)) {
       return false;
@@ -111,13 +110,11 @@ abstract class Relationship<E extends DataModel<E>, N>
     return true;
   }
 
-  @override
   bool contains(Object? element) {
     return _iterable.contains(element);
   }
 
   /// Removes a [value] from this [Relationship]
-  @override
   bool remove(Object? value, {bool notify = true}) {
     assert(value is E);
     final model = value as E;
@@ -135,17 +132,27 @@ abstract class Relationship<E extends DataModel<E>, N>
     return _uninitializedModels.remove(model);
   }
 
-  @override
   Iterator<E> get iterator => _iterable.iterator;
 
-  @override
   E? lookup(Object? element) => lookup(element);
 
-  @override
   Set<E> toSet() => _iterable.toSet();
 
-  @override
+  List<E> toList() => _iterable.toList();
+
   int get length => _iterable.length;
+
+  //
+
+  E get first => _iterable.first;
+
+  bool get isEmpty => _iterable.isEmpty;
+
+  bool get isNotEmpty => _iterable.isNotEmpty;
+
+  Iterable<E> where(bool Function(E) test) => _iterable.where(test);
+
+  Iterable<T> map<T>(T Function(E) f) => _iterable.map(f);
 
   // support methods
 
@@ -200,25 +207,26 @@ abstract class Relationship<E extends DataModel<E>, N>
 
   // equality
 
-  @override
-  bool operator ==(dynamic other) =>
-      identical(this, other) ||
-      other is Relationship &&
-          isInitialized &&
-          other.isInitialized &&
-          _ownerKey == other._ownerKey &&
-          _name == other._name;
+  // @override
+  // bool operator ==(dynamic other) =>
+  //     identical(this, other) ||
+  //     other is Relationship &&
+  //         isInitialized &&
+  //         other.isInitialized &&
+  //         _ownerKey == other._ownerKey &&
+  //         _name == other._name;
+
+  // @override
+  // int get hashCode {
+  //   if (isInitialized) {
+  //     return Object.hash(runtimeType, _ownerKey, _name);
+  //   } else {
+  //     return runtimeType.hashCode;
+  //   }
+  // }
 
   @override
-  int get hashCode {
-    if (isInitialized) {
-      return Object.hash(runtimeType, _ownerKey, _name);
-    } else {
-      return runtimeType.hashCode;
-    }
-  }
-
-  String get _prop => _iterable.map((e) => e.id).join(', ');
+  List<Object?> get props => [isInitialized, _ownerKey, _name];
 
   @override
   void dispose() {