+++ test/_support/book.freezed.dart
@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'book.dart';
@@ -155,13 +156,17 @@ class _$_BookAuthor extends _BookAuthor {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _BookAuthor &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.books, books) || other.books == books));
+            const DeepCollectionEquality().equals(other.id, id) &&
+            const DeepCollectionEquality().equals(other.name, name) &&
+            const DeepCollectionEquality().equals(other.books, books));
   }
 
   @override
-  int get hashCode => Object.hash(runtimeType, id, name, books);
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(id),
+      const DeepCollectionEquality().hash(name),
+      const DeepCollectionEquality().hash(books));
 
   @JsonKey(ignore: true)
   @override
@@ -351,7 +356,7 @@ class _$_Book extends _Book {
   final int id;
   @override
   final String? title;
-  @JsonKey(defaultValue: 0)
+  @JsonKey()
   @override
   final int numberOfSales;
   @override
@@ -368,17 +373,21 @@ class _$_Book extends _Book {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _Book &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.title, title) || other.title == title) &&
-            (identical(other.numberOfSales, numberOfSales) ||
-                other.numberOfSales == numberOfSales) &&
-            (identical(other.originalAuthor, originalAuthor) ||
-                other.originalAuthor == originalAuthor));
+            const DeepCollectionEquality().equals(other.id, id) &&
+            const DeepCollectionEquality().equals(other.title, title) &&
+            const DeepCollectionEquality()
+                .equals(other.numberOfSales, numberOfSales) &&
+            const DeepCollectionEquality()
+                .equals(other.originalAuthor, originalAuthor));
   }
 
   @override
-  int get hashCode =>
-      Object.hash(runtimeType, id, title, numberOfSales, originalAuthor);
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(id),
+      const DeepCollectionEquality().hash(title),
+      const DeepCollectionEquality().hash(numberOfSales),
+      const DeepCollectionEquality().hash(originalAuthor));
 
   @JsonKey(ignore: true)
   @override