+++ test/_support/node.freezed.dart
@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'node.dart';
@@ -185,15 +186,19 @@ class _$_Node extends _Node {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _Node &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.parent, parent) || other.parent == parent) &&
-            (identical(other.children, children) ||
-                other.children == children));
+            const DeepCollectionEquality().equals(other.id, id) &&
+            const DeepCollectionEquality().equals(other.name, name) &&
+            const DeepCollectionEquality().equals(other.parent, parent) &&
+            const DeepCollectionEquality().equals(other.children, children));
   }
 
   @override
-  int get hashCode => Object.hash(runtimeType, id, name, parent, children);
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(id),
+      const DeepCollectionEquality().hash(name),
+      const DeepCollectionEquality().hash(parent),
+      const DeepCollectionEquality().hash(children));
 
   @JsonKey(ignore: true)
   @override