+++ test/model/relationship/relationship_test.dart
@@ -54,12 +54,12 @@ void main() async {
     // once p1 exists
     final p1 = Person(id: '1', name: 'Axl', age: 58).init(container.read);
     // it's automatically wired up
-    expect(f1b.persons, {p1});
+    expect(f1b.persons.toSet(), {p1});
 
     // relationships are omitted - so they remain unchanged
     final f1c = familyRemoteAdapter.localAdapter
         .deserialize({'id': '1', 'surname': 'Rose'}).init(container.read);
-    expect(f1c.persons, {p1});
+    expect(f1c.persons.toSet(), {p1});
     expect(f1c.residence!.value, isNotNull);
 
     final p2 = Person(id: '2', name: 'Brian', age: 55).init(container.read);
@@ -71,7 +71,7 @@ void main() async {
       'persons': [keyFor(p2)]
     }).init(container.read);
     // persons should be exactly equal to p2 (Brian)
-    expect(f1d.persons, {p2});
+    expect(f1d.persons.toSet(), {p2});
     // without directly modifying p2, its family should be automatically updated
     expect(p2.family.value, f1d);
     // and by the same token, p1's family should now be null
@@ -130,7 +130,7 @@ void main() async {
     ).init(container.read);
 
     expect(family.residence!.key, isNotNull);
-    expect(family.persons!.keys.length, 3);
+    expect(family.persons.keys.length, 3);
 
     // associate ids with keys
     graph.getKeyForId('people', '1', keyIfAbsent: 'people#c1c1c1');
@@ -139,7 +139,7 @@ void main() async {
     graph.getKeyForId('houses', '98', keyIfAbsent: 'houses#c98d1b');
 
     // test ids
-    expect(family.persons!.ids, {'1', '2', '3'});
+    expect(family.persons.ids, {'1', '2', '3'});
 
     // (2) then load persons
 
@@ -147,7 +147,7 @@ void main() async {
     Person(id: '2', name: 'z2', age: 33).init(container.read);
 
     // (3) assert two first are linked, third one null, residence is null
-    expect(family.persons!.length, 2);
+    expect(family.persons.length, 2);
     expect(family.residence!.value, isNull);
 
     // (4) load the last person and assert it exists now
@@ -166,23 +166,23 @@ void main() async {
     final igor = Person(name: 'Igor', age: 33).init(container.read);
     final f1 = Family(surname: 'Kamchatka', persons: {igor}.asHasMany)
         .init(container.read);
-    expect(f1.persons!.first.family.value, f1);
+    expect(f1.persons.first.family.value, f1);
 
     final igor1b =
         Person(name: 'Igor', age: 33, family: BelongsTo()).init(container.read);
 
     final f1b = Family(surname: 'Kamchatka', persons: {igor1b}.asHasMany)
         .init(container.read);
-    expect(f1b.persons!.first.family.value!.surname, 'Kamchatka');
+    expect(f1b.persons.first.family.value!.surname, 'Kamchatka');
 
     final f2 =
         Family(surname: 'Kamchatka', persons: HasMany()).init(container.read);
     final igor2 =
         Person(name: 'Igor', age: 33, family: BelongsTo()).init(container.read);
-    f2.persons!.add(igor2);
-    expect(f2.persons!.first.family.value!.surname, 'Kamchatka');
+    f2.persons.add(igor2);
+    expect(f2.persons.first.family.value!.surname, 'Kamchatka');
 
-    f2.persons!.remove(igor2);
+    f2.persons.remove(igor2);
     expect(f2.persons, isEmpty);
 
     final residence =
@@ -205,13 +205,13 @@ void main() async {
     final family =
         Family(id: '229', surname: 'Rose', persons: {brian}.asHasMany)
             .init(container.read);
-    expect(family.persons!.length, 1);
+    expect(family.persons.length, 1);
 
     // new family comes in locally with no persons relationship info
     final family2 = Family(id: '229', surname: 'Rose', persons: HasMany())
         .init(container.read);
     // it should keep the relationships unaltered
-    expect(family2.persons!.length, 1);
+    expect(family2.persons.length, 1);
 
     // new family comes in from API (simulate) with no persons relationship info
     final family3 = familyRemoteAdapter
@@ -219,7 +219,7 @@ void main() async {
         .model!
         .init(container.read);
     // it should keep the relationships unaltered
-    expect(family3.persons!.length, 1);
+    expect(family3.persons.length, 1);
 
     // new family comes in from API (simulate) with empty persons relationship
     final family4 = familyRemoteAdapter
@@ -227,7 +227,7 @@ void main() async {
         .model!
         .init(container.read);
     // it should keep the relationships unaltered
-    expect(family4.persons!.length, 0);
+    expect(family4.persons.length, 0);
 
     // since we're passing a key (not an ID)
     // we MUST use the local adapter serializer
@@ -239,7 +239,7 @@ void main() async {
 
     graph.getKeyForId('people', '231', keyIfAbsent: 'people#231aaa');
     final axl = Person(id: '231', name: 'Axl', age: 58).init(container.read);
-    expect(family5.persons, {axl});
+    expect(family5.persons.toSet(), {axl});
   });
 
   test('scenario #5: one-way relationships', () {
@@ -248,7 +248,7 @@ void main() async {
     final zoe = Dog(name: 'Zoe');
     final f1 = Family(surname: 'Carlson', dogs: {jerry, zoe}.asHasMany)
         .init(container.read);
-    expect(f1.dogs, {jerry, zoe});
+    expect(f1.dogs!.toSet(), {jerry, zoe});
   });
 
   test('self-ref with freezed', () {
@@ -306,12 +306,9 @@ void main() async {
         equals(BookAuthor(id: 15, name: 'Steve-O', books: HasMany({book}))
             .init(container.read)));
 
-    // expect(HasMany({book}), isNot(HasMany<Book>()));
-    // equals() will ignore HasMany's ==/hashCode since it's a Set
-    // it will recursively match the elements
-    expect(HasMany<Book>() == HasMany<Book>(), isFalse);
-    // this will pass as book == book
-    expect(HasMany({book}), equals(HasMany<Book>({book})));
+    // expect a LateInitializationError when trying
+    // to compare uninitialized relationships
+    expect(() => HasMany<Book>() == HasMany<Book>(), throwsA(isA<Error>()));
 
     expect(author.books!.first.originalAuthor.toString(),
         'BelongsTo<BookAuthor>(15)');
@@ -327,7 +324,7 @@ void main() async {
     }.asHasMany;
 
     expect(rel.isNotEmpty, isTrue);
-    expect(rel.toList(), isA<List>());
+    expect(rel.toSet(), isA<Set>());
     expect(rel.where((e) => e.age! > 40), hasLength(1));
     expect(rel.map((e) => e.age! - 10).toSet(), {39, 13});
   });